Move Password Reminder actions.js
