+++ src/components/themed/PasswordReminderModal.js
@@ -7,9 +7,9 @@ import { type AirshipBridge } from 'react-native-airship'
 import { Actions } from 'react-native-router-flux'
 import { connect } from 'react-redux'
 
+import { postponePasswordReminder, requestChangePassword } from '../../actions/PasswordReminderActions.js'
 import { CHANGE_PASSWORD } from '../../constants/indexConstants.js'
 import s from '../../locales/strings.js'
-import { postponePasswordReminder, requestChangePassword } from '../../modules/UI/components/PasswordReminderModal/actions.js'
 import { type Dispatch, type RootState } from '../../types/reduxTypes.js'
 import { showToast } from '../services/AirshipInstance.js'
 import { EdgeTextField } from './EdgeTextField.js'
